code cleanup #98: files renamed
