+++ Notepad2e.vcxproj
@@ -324,7 +324,7 @@
     <ClCompile Include="src\Edit.c" />
     <ClCompile Include="src\Extension\EditHelper.c" />
     <ClCompile Include="src\Extension\EditHelperEx.cpp" />
-    <ClCompile Include="src\Extension\HLSelection.c" />
+    <ClCompile Include="src\Extension\ExtSelection.c" />
     <ClCompile Include="src\Extension\InlineProgressBarCtrl.c" />
     <ClCompile Include="src\Extension\MainWndHelper.c" />
     <ClCompile Include="src\Extension\StrToHex.c" />
@@ -402,19 +402,17 @@
     <ClInclude Include="src\Edit.h" />
     <ClInclude Include="src\Extension\EditHelper.h" />
     <ClInclude Include="src\Extension\EditHelperEx.h" />
-    <ClInclude Include="src\Extension\HLSelection.h" />
+    <ClInclude Include="src\Extension\ExtSelection.h" />
     <ClInclude Include="src\Extension\InlineProgressBarCtrl.h" />
     <ClInclude Include="src\Extension\MainWndHelper.h" />
     <ClInclude Include="src\Extension\SciCall.h" />
+    <ClInclude Include="src\Extension\stdafx.h" />
     <ClInclude Include="src\Extension\StrToHex.h" />
     <ClInclude Include="src\Extension\tinyexpr\tinyexpr.h" />
     <ClInclude Include="src\Extension\Utils.h" />
     <ClInclude Include="src\Helpers.h" />
-    <ClInclude Include="src\HLSelection.h" />
-    <ClInclude Include="src\InlineProgressBarCtrl.h" />
     <ClInclude Include="src\Notepad2.h" />
     <ClInclude Include="src\SciCall.h" />
-    <ClInclude Include="src\StrToHex.h" />
     <ClInclude Include="src\Styles.h" />
     <ClInclude Include="src\resource.h" />
     <ClInclude Include="src\version.h" />
