--- src/Extension/readme.txt
@@ -1,1 +1,0 @@
-temporary file
